refactor: 중복된 클래스를 제거한다.
