+++ src/main/java/com/clova/anifriends/domain/animal/dto/response/FindAnimalsResponse.java
@@ -4,7 +4,6 @@ import com.clova.anifriends.domain.animal.Animal;
 import com.clova.anifriends.domain.animal.repository.response.FindAnimalsResult;
 import com.clova.anifriends.domain.common.PageInfo;
 import java.util.List;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Slice;
 
 public record FindAnimalsResponse(
@@ -40,14 +39,6 @@ public record FindAnimalsResponse(
             );
         }
 
-    }
-
-    public static FindAnimalsResponse from(Page<Animal> pagination) {
-        PageInfo pageInfo = PageInfo.of(pagination.getTotalElements(), pagination.hasNext());
-        List<FindAnimalResponse> findAnimalByVolunteerResponses = pagination.get()
-            .map(FindAnimalResponse::from).toList();
-
-        return new FindAnimalsResponse(pageInfo, findAnimalByVolunteerResponses);
     }
 
     public static FindAnimalsResponse fromV2(Slice<FindAnimalsResult> animalsWithPagination,
